--- src/main/java/account/application/port/in/DeleteAccountUseCase.java
@@ -1,4 +1,0 @@
-package account.application.port.in;
-
-public interface DeleteAccountUseCase {
-}
