--- src/main/java/account/application/port/in/ForceDeleteAccountUseCase.java
@@ -1,4 +1,0 @@
-package account.application.port.in;
-
-public interface ForceDeleteAccountUseCase {
-}
