--- src/main/java/account/domain/model/aggregate/AccountId.java
@@ -1,10 +1,0 @@
-package account.domain.model.aggregate;
-
-public record AccountId(
-        Long id
-) {
-    public static AccountId EMPTY = new AccountId(null);
-    public boolean isNull() {
-      return id == null;
-  }
-}
