--- src/main/java/account/domain/model/aggregate/Email.java
@@ -1,9 +1,0 @@
-package account.domain.model.aggregate;
-
-public record Email(
-        String email
-) {
-    public static Email from(String email) {
-        return new Email(email);
-    }
-}
