--- src/main/java/account/domain/model/aggregate/ProfileNickName.java
@@ -1,9 +1,0 @@
-package account.domain.model.aggregate;
-
-public record ProfileNickName(
-        String nickName
-) {
-    public static ProfileNickName from(String nickName) {
-        return new ProfileNickName(nickName);
-    }
-}
