--- src/main/java/account/domain/model/aggregate/Url.java
@@ -1,9 +1,0 @@
-package account.domain.model.aggregate;
-
-public record Url(
-        String url
-) {
-    public static Url from(String url) {
-        return new Url(url);
-    }
-}
