--- src/main/java/account/infrastructure/adapter/in/rest/mapper/AccountRestModelMapper.java
@@ -1,10 +1,0 @@
-package account.infrastructure.adapter.in.rest.mapper;
-
-import account.domain.model.commands.CreateAccountCommand;
-import account.infrastructure.adapter.in.rest.model.CreateAccountRequestDto;
-import org.mapstruct.Mapper;
-
-@Mapper
-public interface AccountRestModelMapper {
-    CreateAccountCommand toCommand(CreateAccountRequestDto createAccountRequestDto);
-}
