+++ src/main/java/com/studioplayground/azbackend/account/application/port/in/ForceDeleteAccountUseCase.java
@@ -1,0 +1,4 @@
+package com.studioplayground.azbackend.account.application.port.in;
+
+public interface ForceDeleteAccountUseCase {
+}
