+++ src/main/java/com/studioplayground/azbackend/account/application/port/in/SignUpAccountUseCase.java
@@ -1,0 +1,8 @@
+package com.studioplayground.azbackend.account.application.port.in;
+
+import com.studioplayground.azbackend.account.domain.model.aggregate.Account;
+import com.studioplayground.azbackend.account.domain.model.commands.CreateAccountCommand;
+
+public interface SignUpAccountUseCase {
+    Account create(CreateAccountCommand createAccountCommand);
+}
