+++ src/main/java/com/studioplayground/azbackend/account/domain/model/aggregate/Email.java
@@ -1,0 +1,9 @@
+package com.studioplayground.azbackend.account.domain.model.aggregate;
+
+public record Email(
+        String email
+) {
+    public static Email from(String email) {
+        return new Email(email);
+    }
+}
