+++ src/main/java/com/studioplayground/azbackend/account/infrastructure/adapter/in/rest/mapper/AccountRestModelMapper.java
@@ -1,0 +1,10 @@
+package com.studioplayground.azbackend.account.infrastructure.adapter.in.rest.mapper;
+
+import com.studioplayground.azbackend.account.domain.model.commands.CreateAccountCommand;
+import com.studioplayground.azbackend.account.infrastructure.adapter.in.rest.model.CreateAccountRequestDto;
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface AccountRestModelMapper {
+    CreateAccountCommand toCommand(CreateAccountRequestDto createAccountRequestDto);
+}
